+++ standingsrequests/tests/test_managers.py
@@ -88,6 +88,7 @@ class TestContactSetManager(NoSocketsTestCase):
             (1008, -5),
             (1009, -10),
             (1010, 5),
+            (1110, 5.0),
             (2001, 10.0),
             (2003, 5.0),
             (2102, -10.0),